+++ src/rest_toolkit/state.py
@@ -1,12 +1,13 @@
 class RestState(object):
-    def __init__(self, resource_class, route_path):
+    def __init__(self, resource_class, route_path, route_name=None):
         self.resource_class = resource_class
         self.route_path = route_path
         self.views = {}
         self.controllers = {}
-
-    def route_name(self):
-        return 'rest-%s' % self.resource_class.__name__
+        if route_name:
+            self.route_name = route_name
+        else:
+            self.route_name = 'rest-%s' % self.resource_class.__name__
 
     def add_method(self, method, view):
         self.views[method] = view
@@ -15,8 +16,9 @@ class RestState(object):
         self.controllers[name] = view
 
     @classmethod
-    def add_to_resource(cls, resource_class, route_path):
-        resource_class.__rest__ = state = RestState(resource_class, route_path)
+    def add_to_resource(cls, resource_class, route_path, route_name):
+        resource_class.__rest__ = state = RestState(resource_class,
+                route_path, route_name)
         return state
 
     @classmethod